pyproject.toml
[project]
name = "second-prac-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.4.1",
    "openai-agents>=0.0.4",
    "python-dotenv>=1.0.1",
    "tavily-python>=0.5.1",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.8.1",
]
